CodeBlog/requirements.txt
beautifulsoup4==4.6.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-cors-headers==3.1.0
django-modelcluster==4.4
django-pipeline==1.6.14
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.2
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
manifesto==0.5
oauthlib==3.0.2
Pillow==5.4.1
ply==3.11
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
slimit==0.8.1
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.5.1
wagtail-cache==0.5.2
wagtail-modelchooser==2.2.1
wagtail-themes==0.3.0
wagtailcodeblock==1.15.0.0
webencodings==0.5.1
Willow==1.1